Refactor DSFR color utilities and update build outputs
Moved DSFR color utilities from src/utils/dsfr-colors.js to src/fr/colors.js and updated references to use the new fr.colors export. Removed unused typography generation and related build logic. Changed build outputs from .ts to .js files for generated color, spacing, breakpoint, and classNames modules. Updated chartData to use fr.colors. Added type definitions and fr module entry point in src/fr.
