+++ src/components/LineChart.vue
@@ -394,6 +394,7 @@ export default {
       this.colorParse = [];
       this.colorHover = [];
       const palette = this.choosePalette();
+      console.log('Selected Palette:', palette);
       for (let i = 0; i < this.yparse.length; i++) {
         if (this.tmpColorParse[i]) {
           const color = this.tmpColorParse[i];
@@ -427,6 +428,7 @@ export default {
       }
     },
     choosePalette() {
+      console.log("choosePalette", this.selectedPalette);
       // Using the refactored choosePalette function from utils
       return choosePalette(this.selectedPalette);
     },
