+++ src/fr/index.js
@@ -1,0 +1,7 @@
+
+export * from "./colors";
+import { colors } from "./colors";
+
+export const fr = {
+  colors,
+};
